defined models in dbstorage to correct errors

+++ models/engine/db_storage.py
@@ -3,27 +3,25 @@
 Contains the class DBStorage
 """
 
-import models
-from models.amenity import Amenity
-from models.base_model import BaseModel, Base
-from models.city import City
-from models.place import Place
-from models.review import Review
-from models.state import State
-from models.user import User
-from os import getenv
+import os
 import sqlalchemy
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
+from models import Base, Amenity, City, Place, Review, State, User
 
+# Assuming 'models' is a package that contains your ORM model definitions
+from models import storage
+
+# Dictionary mapping class names to classes
 classes = {
-    "Amenity": Amenity, 
+    "Amenity": Amenity,
     "City": City,
-    "Place": Place, 
-    "Review": Review, 
-    "State": State, 
+    "Place": Place,
+    "Review": Review,
+    "State": State,
     "User": User
 }
+
 
 class DBStorage:
     """Interacts with the MySQL database"""
@@ -32,14 +30,17 @@ class DBStorage:
 
     def __init__(self):
         """Instantiate a DBStorage object"""
-        HBNB_MYSQL_USER = getenv('HBNB_MYSQL_USER')
-        HBNB_MYSQL_PWD = getenv('HBNB_MYSQL_PWD')
-        HBNB_MYSQL_HOST = getenv('HBNB_MYSQL_HOST')
-        HBNB_MYSQL_DB = getenv('HBNB_MYSQL_DB')
-        HBNB_ENV = getenv('HBNB_ENV')
+        HBNB_MYSQL_USER = os.getenv('HBNB_MYSQL_USER')
+        HBNB_MYSQL_PWD = os.getenv('HBNB_MYSQL_PWD')
+        HBNB_MYSQL_HOST = os.getenv('HBNB_MYSQL_HOST')
+        HBNB_MYSQL_DB = os.getenv('HBNB_MYSQL_DB')
+        HBNB_ENV = os.getenv('HBNB_ENV')
 
-        self.__engine = create_engine(
-            'mysql+pymysql://{}:{}@{}/{}'.format(HBNB_MYSQL_USER, HBNB_MYSQL_PWD, HBNB_MYSQL_HOST, HBNB_MYSQL_DB))
+        if not all([HBNB_MYSQL_USER, HBNB_MYSQL_PWD, HBNB_MYSQL_HOST, HBNB_MYSQL_DB]):
+            raise EnvironmentError("Database environment variables not set")
+
+        conn_str = f'mysql+pymysql://{HBNB_MYSQL_USER}:{HBNB_MYSQL_PWD}@{HBNB_MYSQL_HOST}/{HBNB_MYSQL_DB}'
+        self.__engine = create_engine(conn_str)
 
         if HBNB_ENV == "test":
             Base.metadata.drop_all(self.__engine)
@@ -51,7 +52,7 @@ class DBStorage:
             if cls is None or cls is classes[clss] or cls is clss:
                 objs = self.__session.query(classes[clss]).all()
                 for obj in objs:
-                    key = obj.__class__.__name__ + '.' + obj.id
+                    key = obj.__class__.__name__ + '.' + str(obj.id)
                     new_dict[key] = obj
         return new_dict
 
@@ -73,7 +74,7 @@ class DBStorage:
         Base.metadata.create_all(self.__engine)
         sess_factory = sessionmaker(bind=self.__engine, expire_on_commit=False)
         Session = scoped_session(sess_factory)
-        self.__session = Session
+        self.__session = Session()
 
     def close(self):
         """Call remove() method on the private session attribute"""
@@ -87,9 +88,9 @@ class DBStorage:
         if cls not in classes.values():
             return None
 
-        all_cls = models.storage.all(cls)
+        all_cls = storage.all(cls)
         for value in all_cls.values():
-            if (value.id == id):
+            if str(value.id) == id:
                 return value
 
         return None
@@ -103,9 +104,9 @@ class DBStorage:
         if not cls:
             count = 0
             for clas in all_class:
-                count += len(models.storage.all(clas).values())
+                count += len(storage.all(clas).values())
         else:
-            count = len(models.storage.all(cls).values())
+            count = len(storage.all(cls).values())
 
         return count
